Moved autocomplete to own component
